feat: list & block online users

+++ model/server_api.go
@@ -16,7 +16,7 @@ type StreamServer struct {
 
 type StreamServerData struct {
 	Now     int64          `json:"now,omitempty"`
-	Online  uint64         `json:"online,omitempty"`
+	Online  int            `json:"online,omitempty"`
 	Servers []StreamServer `json:"servers,omitempty"`
 }
